Added more to checkout process

+++ app/controllers/orders_controller.rb
@@ -1,19 +1,22 @@
 class OrdersController < ApplicationController
   def checkout
-    customer = current_customer
-    province = customer&.province
+    if current_customer
+      province = current_customer.province
 
-    unless province
-      redirect_to root_path, alert: "We couldn't determine your province. Please update your profile." and return
+      unless province
+        redirect_to root_path, alert: "We couldn't determine your province. Please update your profile." and return
+      end
+
+      @province_name = province.name
+      @gst_rate = province.gst_rate.to_f / 100
+      @pst_rate = province.pst_rate.to_f / 100
+      @hst_rate = province.hst_rate.to_f / 100
+    else
+      @province_name = "Unknown"
+      @gst_rate = @pst_rate = @hst_rate = 0
     end
-    @province_name = province.name
-    @gst_rate = province.gst_rate.to_f / 100
-    @pst_rate = province.pst_rate.to_f / 100
-    @hst_rate = province.hst_rate.to_f / 100
-
 
     tax_multiplier = 1 + @gst_rate + @pst_rate + @hst_rate
-
     @cart_items = Product.find(cart.keys)
 
     @cart_totals = @cart_items.each_with_object({}) do |product, hash|
@@ -23,28 +26,54 @@ class OrdersController < ApplicationController
     end
 
     @cart_total = @cart_totals.values.sum
-    @order = Order.new
   end
 
-
-
   def start_payment
-  session = Stripe::Checkout::Session.create(
-    payment_method_types: ["card"],
-    line_items: build_line_items,
-    mode: "payment",
-    success_url: order_success_url,
-    cancel_url: checkout_url
-  )
+    session = Stripe::Checkout::Session.create(
+      payment_method_types: ["card"],
+      line_items: build_line_items,
+      mode: "payment",
+      success_url: order_success_url,
+      cancel_url: checkout_url
+    )
 
-  redirect_to session.url, allow_other_host: true
+    redirect_to session.url, allow_other_host: true
   end
 
   def success
     @order = Order.find_by(id: session[:order_id])
-    session.delete(:cart)
-  end
 
+    if @order
+      province = @order.province
+      gst_rate = province.gst_rate.to_f / 100
+      pst_rate = province.pst_rate.to_f / 100
+      hst_rate = province.hst_rate.to_f / 100
+      tax_multiplier = 1 + gst_rate + pst_rate + hst_rate
+
+      cart_items = Product.find(cart.keys)
+      cart_totals = cart_items.each_with_object({}) do |product, hash|
+        quantity = cart[product.id.to_s].to_i
+        price_with_tax = product.current_price * quantity * tax_multiplier
+        hash[product.id] = price_with_tax
+      end
+
+      cart_total = cart_totals.values.sum
+
+      @order.update(
+        subtotal: cart_total,
+        gst_amount: cart_total * gst_rate,
+        pst_amount: cart_total * pst_rate,
+        hst_amount: cart_total * hst_rate,
+        total_price: cart_total * tax_multiplier,
+        order_status: "paid"
+      )
+
+      session.delete(:order_id)
+      session.delete(:cart)
+    else
+      redirect_to checkout_url, alert: "Order not found. Please try again."
+    end
+  end
 
   def show
     @order = Order.find(params[:id])
@@ -66,6 +95,4 @@ class OrdersController < ApplicationController
       }
     end
   end
-
-
 end
